Isolate rhpv3 code in v3 client
