+++ worker/download.go
@@ -13,6 +13,7 @@ import (
 	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
 	"go.sia.tech/renterd/api"
+	rhp3 "go.sia.tech/renterd/internal/rhp/v3"
 	"go.sia.tech/renterd/internal/utils"
 	"go.sia.tech/renterd/object"
 	"go.uber.org/zap"
@@ -757,11 +758,11 @@ loop:
 				}
 
 				// handle lost sectors
-				if isSectorNotFound(resp.err) {
+				if rhp3.IsSectorNotFound(resp.err) {
 					if err := s.mgr.os.DeleteHostSector(ctx, resp.req.host.PublicKey(), resp.req.root); err != nil {
 						s.mgr.logger.Errorw("failed to mark sector as lost", "hk", resp.req.host.PublicKey(), "root", resp.req.root, zap.Error(err))
 					}
-				} else if isPriceTableGouging(resp.err) && s.overpay && !resp.req.overpay {
+				} else if rhp3.IsPriceTableGouging(resp.err) && s.overpay && !resp.req.overpay {
 					resp.req.overpay = true // ensures we don't retry the same request over and over again
 					gouging = append(gouging, resp.req)
 				}
