+++ worker/host.go
@@ -2,10 +2,8 @@ package worker
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io"
-	"math"
 	"time"
 
 	rhpv2 "go.sia.tech/core/rhp/v2"
@@ -13,11 +11,8 @@ import (
 	"go.sia.tech/core/types"
 	"go.sia.tech/renterd/api"
 	"go.sia.tech/renterd/internal/gouging"
+	rhp3 "go.sia.tech/renterd/internal/rhp/v3"
 	"go.uber.org/zap"
-)
-
-var (
-	errFailedToCreatePayment = errors.New("failed to create payment")
 )
 
 type (
@@ -27,7 +22,8 @@ type (
 		DownloadSector(ctx context.Context, w io.Writer, root types.Hash256, offset, length uint32, overpay bool) error
 		UploadSector(ctx context.Context, sectorRoot types.Hash256, sector *[rhpv2.SectorSize]byte, rev types.FileContractRevision) error
 
-		FetchPriceTable(ctx context.Context, rev *types.FileContractRevision) (hpt api.HostPriceTable, err error)
+		PriceTable(ctx context.Context, rev *types.FileContractRevision) (hpt api.HostPriceTable, err error)
+		PriceTableUnpaid(ctx context.Context) (hpt api.HostPriceTable, err error)
 		FetchRevision(ctx context.Context, fetchTimeout time.Duration) (types.FileContractRevision, error)
 
 		FundAccount(ctx context.Context, balance types.Currency, rev *types.FileContractRevision) error
@@ -50,10 +46,10 @@ type (
 		siamuxAddr string
 
 		acc                      *account
+		client                   *rhp3.Client
 		bus                      Bus
 		contractSpendingRecorder ContractSpendingRecorder
 		logger                   *zap.SugaredLogger
-		transportPool            *transportPoolV3
 		priceTables              *priceTables
 	}
 )
@@ -65,6 +61,7 @@ var (
 
 func (w *Worker) Host(hk types.PublicKey, fcid types.FileContractID, siamuxAddr string) Host {
 	return &host{
+		client:                   w.rhp3Client,
 		hk:                       hk,
 		acc:                      w.accounts.ForHost(hk),
 		bus:                      w.bus,
@@ -74,7 +71,6 @@ func (w *Worker) Host(hk types.PublicKey, fcid types.FileContractID, siamuxAddr 
 		siamuxAddr:               siamuxAddr,
 		renterKey:                w.deriveRenterKey(hk),
 		accountKey:               w.accounts.deriveAccountKey(hk),
-		transportPool:            w.transportPoolV3,
 		priceTables:              w.priceTables,
 	}
 }
@@ -97,30 +93,8 @@ func (h *host) DownloadSector(ctx context.Context, w io.Writer, root types.Hash2
 		return fmt.Errorf("%w: %v", gouging.ErrPriceTableGouging, breakdown.DownloadErr)
 	}
 
-	// return errBalanceInsufficient if balance insufficient
-	defer func() {
-		if isBalanceInsufficient(err) {
-			err = fmt.Errorf("%w %v, err: %v", errBalanceInsufficient, h.hk, err)
-		}
-	}()
-
 	return h.acc.WithWithdrawal(ctx, func() (amount types.Currency, err error) {
-		err = h.transportPool.withTransportV3(ctx, h.hk, h.siamuxAddr, func(ctx context.Context, t *transportV3) error {
-			cost, err := readSectorCost(hpt, uint64(length))
-			if err != nil {
-				return err
-			}
-
-			payment := rhpv3.PayByEphemeralAccount(h.acc.id, cost, pt.HostBlockHeight+defaultWithdrawalExpiryBlocks, h.accountKey)
-			cost, refund, err := RPCReadSector(ctx, t, w, hpt, &payment, offset, length, root)
-			if err != nil {
-				return err
-			}
-
-			amount = cost.Sub(refund)
-			return nil
-		})
-		return
+		return h.client.ReadSector(ctx, offset, length, root, w, h.hk, h.siamuxAddr, h.acc.id, h.accountKey, hpt)
 	})
 }
 
@@ -130,88 +104,81 @@ func (h *host) UploadSector(ctx context.Context, sectorRoot types.Hash256, secto
 	if err != nil {
 		return err
 	}
-
-	// prepare payment
-	//
-	// TODO: change to account payments once we have the means to check for an
-	// insufficient balance error
-	expectedCost, _, _, err := uploadSectorCost(pt, rev.WindowEnd)
+	// upload
+	cost, err := h.client.AppendSector(ctx, sectorRoot, sector, &rev, h.hk, h.siamuxAddr, h.acc.id, pt, h.renterKey)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to upload sector: %w", err)
 	}
-	if rev.RevisionNumber == math.MaxUint64 {
-		return fmt.Errorf("revision number has reached max, fcid %v", rev.ParentID)
-	}
-	payment, ok := rhpv3.PayByContract(&rev, expectedCost, h.acc.id, h.renterKey)
-	if !ok {
-		return errFailedToCreatePayment
-	}
-
-	var cost types.Currency
-	err = h.transportPool.withTransportV3(ctx, h.hk, h.siamuxAddr, func(ctx context.Context, t *transportV3) error {
-		cost, err = RPCAppendSector(ctx, t, h.renterKey, pt, &rev, &payment, sectorRoot, sector)
-		return err
-	})
-	if err != nil {
-		return err
-	}
-
 	// record spending
 	h.contractSpendingRecorder.Record(rev, api.ContractSpending{Uploads: cost})
 	return nil
 }
 
 func (h *host) RenewContract(ctx context.Context, rrr api.RHPRenewRequest) (_ rhpv2.ContractRevision, _ []types.Transaction, _, _ types.Currency, err error) {
-	// Try to get a valid pricetable.
-	ptCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-	var pt *rhpv3.HostPriceTable
-	hpt, err := h.priceTables.fetch(ptCtx, h.hk, nil)
-	if err == nil {
-		pt = &hpt.HostPriceTable
-	} else {
-		h.logger.Infof("unable to fetch price table for renew: %v", err)
+	gc, err := h.gougingChecker(ctx, false)
+	if err != nil {
+		return rhpv2.ContractRevision{}, nil, types.ZeroCurrency, types.ZeroCurrency, err
+	}
+	revision, err := h.client.Revision(ctx, h.fcid, h.hk, h.siamuxAddr)
+	if err != nil {
+		return rhpv2.ContractRevision{}, nil, types.ZeroCurrency, types.ZeroCurrency, err
 	}
 
-	var contractPrice types.Currency
-	var rev rhpv2.ContractRevision
-	var txnSet []types.Transaction
-	var renewErr error
-	var fundAmount types.Currency
-	err = h.transportPool.withTransportV3(ctx, h.hk, h.siamuxAddr, func(ctx context.Context, t *transportV3) (err error) {
-		// NOTE: to avoid an edge case where the contract is drained and can
-		// therefore not be used to pay for the revision, we simply don't pay
-		// for it.
-		_, err = RPCLatestRevision(ctx, t, h.fcid, func(revision *types.FileContractRevision) (rhpv3.HostPriceTable, rhpv3.PaymentMethod, error) {
-			// Renew contract.
-			rev, txnSet, contractPrice, fundAmount, renewErr = RPCRenew(ctx, rrr, h.bus, t, pt, *revision, h.renterKey, h.logger)
-			return rhpv3.HostPriceTable{}, nil, nil
-		})
-		return err
-	})
+	// Renew contract.
+	rev, txnSet, contractPrice, fundAmount, err := h.client.Renew(ctx, rrr, gc, func(ctx context.Context, revision types.FileContractRevision, hostAddress, renterAddress types.Address, renterKey types.PrivateKey, renterFunds, minNewCollateral, maxFundAmount types.Currency, pt rhpv3.HostPriceTable, endHeight, windowSize, expectedStorage uint64) (api.WalletPrepareRenewResponse, func(context.Context, types.Transaction, *error), error) {
+		// prepare renew
+		resp, err := h.bus.WalletPrepareRenew(ctx, revision, hostAddress, renterAddress, renterKey, renterFunds, minNewCollateral, maxFundAmount, pt, endHeight, windowSize, expectedStorage)
+		if err != nil {
+			return api.WalletPrepareRenewResponse{}, h.discardTxnOnError, err
+		}
+		return resp, nil, nil
+	}, func(ctx context.Context, txn *types.Transaction, toSign []types.Hash256, cf types.CoveredFields) error {
+		// sign txn
+		return h.bus.WalletSign(ctx, txn, toSign, cf)
+	}, revision, h.renterKey)
 	if err != nil {
 		return rhpv2.ContractRevision{}, nil, contractPrice, fundAmount, err
 	}
-	return rev, txnSet, contractPrice, fundAmount, renewErr
+	return rev, txnSet, contractPrice, fundAmount, err
 }
 
-func (h *host) FetchPriceTable(ctx context.Context, rev *types.FileContractRevision) (hpt api.HostPriceTable, err error) {
+func (h *host) discardTxnOnError(ctx context.Context, txn types.Transaction, err *error) {
+	if *err != nil {
+		h.logger.Error("discarding transaction due to error", zap.Error(*err))
+		if err := h.bus.WalletDiscard(ctx, txn); err != nil {
+			h.logger.Error("failed to discard transaction", zap.Error(err))
+		}
+	}
+}
+
+func (h *host) PriceTableUnpaid(ctx context.Context) (api.HostPriceTable, error) {
+	return h.client.PriceTableUnpaid(ctx, h.hk, h.siamuxAddr)
+}
+
+func (h *host) PriceTable(ctx context.Context, rev *types.FileContractRevision) (hpt api.HostPriceTable, err error) {
 	// fetchPT is a helper function that performs the RPC given a payment function
-	fetchPT := func(paymentFn PriceTablePaymentFunc) (hpt api.HostPriceTable, err error) {
-		err = h.transportPool.withTransportV3(ctx, h.hk, h.siamuxAddr, func(ctx context.Context, t *transportV3) (err error) {
-			hpt, err = RPCPriceTable(ctx, t, paymentFn)
-			return
-		})
-		return
+	fetchPT := func(paymentFn rhp3.PriceTablePaymentFunc) (api.HostPriceTable, error) {
+		return h.client.PriceTable(ctx, h.hk, h.siamuxAddr, paymentFn)
 	}
 
 	// pay by contract if a revision is given
 	if rev != nil {
-		return fetchPT(h.preparePriceTableContractPayment(rev))
+		return fetchPT(rhp3.PreparePriceTableContractPayment(rev, h.acc.id, h.renterKey))
 	}
 
 	// pay by account
-	return fetchPT(h.preparePriceTableAccountPayment())
+	return fetchPT(rhp3.PreparePriceTableAccountPayment(h.accountKey))
+}
+
+// FetchRevision tries to fetch a contract revision from the host.
+func (h *host) FetchRevision(ctx context.Context, fetchTimeout time.Duration) (types.FileContractRevision, error) {
+	if fetchTimeout > 0 {
+		var cancel context.CancelFunc
+		ctx, cancel = context.WithTimeout(ctx, fetchTimeout)
+		defer cancel()
+	}
+	// Try to fetch the revision with an account first.
+	return h.client.Revision(ctx, h.fcid, h.hk, h.siamuxAddr)
 }
 
 func (h *host) FundAccount(ctx context.Context, balance types.Currency, rev *types.FileContractRevision) error {
@@ -228,51 +195,28 @@ func (h *host) FundAccount(ctx context.Context, balance types.Currency, rev *typ
 	deposit := balance.Sub(curr)
 
 	return h.acc.WithDeposit(ctx, func() (types.Currency, error) {
-		if err := h.transportPool.withTransportV3(ctx, h.hk, h.siamuxAddr, func(ctx context.Context, t *transportV3) error {
-			// fetch pricetable directly to bypass the gouging check
-			pt, err := h.priceTables.fetch(ctx, h.hk, rev)
-			if err != nil {
-				return err
-			}
+		// check whether we have money left in the contract
+		cost := types.NewCurrency64(1)
+		if cost.Cmp(rev.ValidRenterPayout()) >= 0 {
+			return types.ZeroCurrency, fmt.Errorf("insufficient funds to fund account: %v <= %v", rev.ValidRenterPayout(), cost)
+		}
+		availableFunds := rev.ValidRenterPayout().Sub(cost)
 
-			// check only the unused defaults
-			gc, err := GougingCheckerFromContext(ctx, false)
-			if err != nil {
-				return err
-			} else if err := gc.CheckUnusedDefaults(pt.HostPriceTable); err != nil {
-				return fmt.Errorf("%w: %v", gouging.ErrPriceTableGouging, err)
-			}
-
-			// check whether we have money left in the contract
-			cost := types.NewCurrency64(1)
-			if cost.Cmp(rev.ValidRenterPayout()) >= 0 {
-				return fmt.Errorf("insufficient funds to fund account: %v <= %v", rev.ValidRenterPayout(), cost)
-			}
-			availableFunds := rev.ValidRenterPayout().Sub(cost)
-
-			// cap the deposit amount by the money that's left in the contract
-			if deposit.Cmp(availableFunds) > 0 {
-				deposit = availableFunds
-			}
-
-			// create the payment
-			amount := deposit.Add(cost)
-			payment, err := payByContract(rev, amount, rhpv3.Account{}, h.renterKey) // no account needed for funding
-			if err != nil {
-				return err
-			}
-
-			// fund the account
-			if err := RPCFundAccount(ctx, t, &payment, h.acc.id, pt.UID); err != nil {
-				return fmt.Errorf("failed to fund account with %v (excluding cost %v);%w", deposit, cost, err)
-			}
-
-			// record the spend
-			h.contractSpendingRecorder.Record(*rev, api.ContractSpending{FundAccount: amount})
-			return nil
-		}); err != nil {
+		// cap the deposit amount by the money that's left in the contract
+		if deposit.Cmp(availableFunds) > 0 {
+			deposit = availableFunds
+		}
+		// fetch pricetable directly to bypass the gouging check
+		pt, err := h.priceTables.fetch(ctx, h.hk, rev)
+		if err != nil {
 			return types.ZeroCurrency, err
 		}
+		amount := deposit.Add(cost)
+		if err := h.client.FundAccount(ctx, rev, h.hk, h.siamuxAddr, amount, h.acc.id, pt.UID, h.renterKey); err != nil {
+			return types.ZeroCurrency, fmt.Errorf("failed to fund account with %v; %w", deposit, err)
+		}
+		// record the spend
+		h.contractSpendingRecorder.Record(*rev, api.ContractSpending{FundAccount: amount})
 		return deposit, nil
 	})
 }
@@ -293,45 +237,32 @@ func (h *host) SyncAccount(ctx context.Context, rev *types.FileContractRevision)
 	}
 
 	return h.acc.WithSync(ctx, func() (types.Currency, error) {
-		var balance types.Currency
-		err := h.transportPool.withTransportV3(ctx, h.hk, h.siamuxAddr, func(ctx context.Context, t *transportV3) error {
-			payment, err := payByContract(rev, types.NewCurrency64(1), h.acc.id, h.renterKey)
-			if err != nil {
-				return err
-			}
-			balance, err = RPCAccountBalance(ctx, t, &payment, h.acc.id, pt.UID)
-			return err
-		})
-		return balance, err
+		return h.client.SyncAccount(ctx, rev, h.hk, h.siamuxAddr, h.acc.id, pt.UID, h.renterKey)
 	})
 }
 
-// preparePriceTableAccountPayment prepare a payment function to pay for a price
-// table from the given host using the provided revision.
-//
-// NOTE: This is the preferred way of paying for a price table since it is
-// faster and doesn't require locking a contract.
-func (h *host) preparePriceTableAccountPayment() PriceTablePaymentFunc {
-	return func(pt rhpv3.HostPriceTable) (rhpv3.PaymentMethod, error) {
-		account := rhpv3.Account(h.accountKey.PublicKey())
-		payment := rhpv3.PayByEphemeralAccount(account, pt.UpdatePriceTableCost, pt.HostBlockHeight+defaultWithdrawalExpiryBlocks, h.accountKey)
-		return &payment, nil
+func (h *host) gougingChecker(ctx context.Context, criticalMigration bool) (gouging.Checker, error) {
+	gp, err := h.bus.GougingParams(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get gouging params: %w", err)
 	}
+	return newGougingChecker(gp.GougingSettings, gp.ConsensusState, gp.TransactionFee, criticalMigration), nil
 }
 
-// preparePriceTableContractPayment prepare a payment function to pay for a
-// price table from the given host using the provided revision.
-//
-// NOTE: This way of paying for a price table should only be used if payment by
-// EA is not possible or if we already need a contract revision anyway. e.g.
-// funding an EA.
-func (h *host) preparePriceTableContractPayment(rev *types.FileContractRevision) PriceTablePaymentFunc {
-	return func(pt rhpv3.HostPriceTable) (rhpv3.PaymentMethod, error) {
-		refundAccount := rhpv3.Account(h.accountKey.PublicKey())
-		payment, err := payByContract(rev, pt.UpdatePriceTableCost, refundAccount, h.renterKey)
-		if err != nil {
-			return nil, err
-		}
-		return &payment, nil
+// priceTable fetches a price table from the host. If a revision is provided, it
+// will be used to pay for the price table. The returned price table is
+// guaranteed to be safe to use.
+func (h *host) priceTable(ctx context.Context, rev *types.FileContractRevision) (rhpv3.HostPriceTable, error) {
+	pt, err := h.priceTables.fetch(ctx, h.hk, rev)
+	if err != nil {
+		return rhpv3.HostPriceTable{}, err
 	}
+	gc, err := GougingCheckerFromContext(ctx, false)
+	if err != nil {
+		return rhpv3.HostPriceTable{}, err
+	}
+	if breakdown := gc.Check(nil, &pt.HostPriceTable); breakdown.Gouging() {
+		return rhpv3.HostPriceTable{}, fmt.Errorf("%w: %v", gouging.ErrPriceTableGouging, breakdown)
+	}
+	return pt.HostPriceTable, nil
 }
